requirements.txt
streamlit
python-dotenv
serpapi
google-generativeai
langchain
langchain-core
langchain-google-genai
groq
pydantic
scikit-learn
sentence-transformers
numpy